Moved AnyCodable to Commons
