--- Tests/WalletConnectSignTests/Helpers/HelperTypes.swift
@@ -1,10 +1,0 @@
-struct AnyError: Error {}
-
-struct EmptyCodable: Codable {}
-
-struct FailableCodable: Codable {
-    
-    func encode(to encoder: Encoder) throws {
-        throw AnyError()
-    }
-}
